statsd: support sending counters and gauges with timestamp.
This PR adds the `timestamp` parameter to `Gauge()` and `Counter()`.

Metrics sent with a `timestamp` to the DogStatsD server will be directly sent to the intake with no aggregation done server-side.

Even if the client side aggregation is enabled, metrics submitted with a timestamp won't go through any aggregation, they will be written in the serialization buffer as-is.